Refactor "Pattern.Match" into stand-alone "PatternMatcher"
