+++ src/DocoptNet/LeafPattern.cs
@@ -38,46 +38,6 @@ namespace DocoptNet
             return new Pattern[] {};
         }
 
-        public virtual (int Index, LeafPattern Match) SingleMatch(IList<LeafPattern> patterns)
-        {
-            return default;
-        }
-
-        public override MatchResult Match(IList<LeafPattern> left,
-                                          IEnumerable<LeafPattern> collected = null)
-        {
-            var coll = collected ?? new List<LeafPattern>();
-            var (index, match) = SingleMatch(left);
-            if (match == null)
-            {
-                return new MatchResult(false, left, coll);
-            }
-            var left_ = new List<LeafPattern>();
-            left_.AddRange(left.Take(index));
-            left_.AddRange(left.Skip(index + 1));
-            var sameName = coll.Where(a => a.Name == Name).ToList();
-            if (Value != null && (Value.IsList || Value.IsOfTypeInt))
-            {
-                var increment = new ValueObject(1);
-                if (!Value.IsOfTypeInt)
-                {
-                    increment = match.Value.IsString ? new ValueObject(new [] {match.Value})  : match.Value;
-                }
-                if (sameName.Count == 0)
-                {
-                    match.Value = increment;
-                    var res = new List<LeafPattern>(coll) {match};
-                    return new MatchResult(true, left_, res);
-                }
-                sameName[0].Value.Add(increment);
-                return new MatchResult(true, left_, coll);
-            }
-            var resColl = new List<LeafPattern>();
-            resColl.AddRange(coll);
-            resColl.Add(match);
-            return new MatchResult(true, left_, resColl);
-        }
-
         public override string ToString()
         {
             return string.Format("{0}({1}, {2})", GetType().Name, Name, Value);
