+++ src/DocoptNet/OneOrMore.cs
@@ -1,39 +1,10 @@
 namespace DocoptNet
 {
-    using System.Collections.Generic;
-    using System.Diagnostics;
-
     class OneOrMore : BranchPattern
     {
         public OneOrMore(params Pattern[] patterns)
             : base(patterns)
         {
-        }
-
-        public override MatchResult Match(IList<LeafPattern> left,
-                                          IEnumerable<LeafPattern> collected = null)
-        {
-            Debug.Assert(Children.Count == 1);
-            var coll = collected ?? new List<LeafPattern>();
-            var l = left;
-            var c = coll;
-            IList<LeafPattern> l_ = null;
-            var matched = true;
-            var times = 0;
-            while (matched)
-            {
-                // could it be that something didn't match but changed l or c?
-                (matched, l, c) = Children[0].Match(l, c);
-                times += matched ? 1 : 0;
-                if (l_ != null && l_.Equals(l))
-                    break;
-                l_ = l;
-            }
-            if (times >= 1)
-            {
-                return new MatchResult(true, l, c);
-            }
-            return new MatchResult(false, left, coll);
         }
     }
 }
