+++ tests/DocoptNet.Tests/Extensions.cs
@@ -4,7 +4,7 @@ namespace DocoptNet.Tests
     {
         public static MatchResult Match(this Pattern pattern, params LeafPattern[] left)
         {
-            return pattern.Match(left);
+            return pattern.Match(left, null);
         }
     }
 }
